pyproject.toml
[project]
name = "confocal-processing"
version = "0.1.0"
description = "Functions to process confocal imaging"
authors = [
    {name = "Andrea"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "napari (>=0.6.6,<0.7.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "locomyco-imganalysis @ git+https://github.com/labudzki/LOCOMYCO_ImgAnalysis.git",
    "matplotlib (>=3.10.7,<4.0.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlrd (>=2.0.2,<3.0.0)",
    "magicgui (>=0.10.1,<0.11.0)",
    "ipykernel (>=6.29.0,<7.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "pyqt6 (>=6.10.2,<7.0.0)"
   
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

# [tool.poetry.dependencies]
# LOCOMYCO_ImgAnalysis = { git = "https://github.com/labudzki/LOCOMYCO_ImgAnalysis.git", branch = "main" }

